removed password package, added jwt package, removed unnecessary migration, changed relation models id to unsigned big int, role_user seed

+++ database/migrations/2020_02_23_062030_create_category_item_table.php
@@ -15,8 +15,8 @@ class CreateCategoryItemTable extends Migration
     {
         Schema::create('category_item', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('category_id');
-            $table->integer('item_id');
+            $table->unsignedBigInteger('category_id');
+            $table->unsignedBigInteger('item_id');
             $table->timestamps();
         });
     }
